chore: move UserStatistics to a independent component
